removed: navigation folder and its contents moved out
