+++ app/src/main/java/dev/enesky/doodle/app/navigation/DoodleNavHost.kt
@@ -7,7 +7,7 @@ import androidx.navigation.compose.NavHost
 import dev.enesky.core.navigation.DoodleNavigationDestination
 import dev.enesky.feature.details.navigation.DetailsDestination
 import dev.enesky.feature.details.navigation.detailsGraph
-import dev.enesky.feature.login.navigation.loginGraph
+import dev.enesky.feature.login.loginGraph
 import dev.enesky.feature.main.navigation.HomeDestination
 import dev.enesky.feature.main.navigation.homeGraph
