+++ app/src/main/java/dev/enesky/doodle/app/ui/DoodleAppState.kt
@@ -10,7 +10,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import dev.enesky.core.navigation.DoodleNavigationDestination
-import dev.enesky.feature.login.navigation.LoginDestination
+import dev.enesky.feature.login.LoginDestination
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.update
