+++ feature/login/src/main/java/dev/enesky/feature/login/LoginDestination.kt
@@ -1,9 +1,8 @@
-package dev.enesky.feature.login.navigation
+package dev.enesky.feature.login
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import dev.enesky.core.navigation.DoodleNavigationDestination
-import dev.enesky.feature.login.LoginRoute
 
 /**
  * Created by Enes Kamil YILMAZ on 10/11/2023
@@ -17,7 +16,7 @@ object LoginDestination : DoodleNavigationDestination {
 fun NavGraphBuilder.loginGraph(
     onNavigateToHomeDestination: () -> Unit,
 ) = composable(route = LoginDestination.route) {
-    LoginRoute(
+    LoginScreenRoute(
         onNavigateHomeClick = onNavigateToHomeDestination,
     )
 }
